Main module - client details (wip).
git-svn-id: http://svn.apollosoft.net.au:18080/svn/myplanner/trunk@58 65940d0d-0698-491e-b608-6f35d31c8762

+++ core/src/main/java/com/argus/financials/dao/ClientDao.java
@@ -18,7 +18,7 @@ public interface ClientDao extends BaseDAO
      * @param clientId
      * @return
      */
-    Client findById(Integer clientId);
+    Client findById(Long clientId);
 
     /**
      *
